Rename script and pass in FIRRTL file
